update sciprts a bit

+++ run_common_tests.sh
@@ -3,16 +3,17 @@
 
 # -- You need to install this helm plugin
 # helm plugin install https://github.com/quintush/helm-unittest
-curr_dir=${pwd}
 
 common_test_path="library/common-test"
 
-echo "Cleaning old built charts..."
-rm -r "$common_test_path/charts"
+if [ -d "$common_test_path/charts" ]; then
+  echo "Cleaning old built charts..."
+  rm -r "$common_test_path/charts"
+  rm  "$common_test_path/Chart.lock"
+fi
 
 echo "Building common..."
 helm dependency update "$common_test_path"
+
 echo "Running tests..."
-cd "$common_test_path"
-helm unittest --helm3 -f "tests/*/*.yaml" .
-cd "$curr_dir"
+helm unittest --update-snapshot --helm3 -f "tests/*/*.yaml" "./$common_test_path"
